automation/requirements.txt
google-generativeai
schedule
gitpython
requests
pyyaml
Pillow
trafilatura
